requirements.txt
azure-core==1.2.2
azure-functions==1.2.0
azure-storage-blob==12.2.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cryptography==2.8
idna==2.9
isodate==0.6.0
msrest==0.6.11
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
xlrd==1.2.0
